requirements.txt
# --- Dashboard & UI ---
streamlit>=1.25.0
streamlit-option-menu==0.4.0
streamlit-aggrid>=0.5.16

# --- Data handling ---
pandas>=2.0.0
numpy>=1.25.0
python-dateutil

# --- Visualization ---
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.12.2
folium>=0.15.1
streamlit-folium>=0.11.0

# --- Supabase / API ---
supabase>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# --- Machine Learning / Model ---
scikit-learn>=1.2.2
joblib>=1.3.0
xgboost>=1.7.6
lightgbm>=4.4.2
tensorflow>=2.15.0
keras>=2.15.0

# --- Utilities / Caching ---
pydantic>=2.5.1
tqdm>=4.66.0

# --- Geospatial / Map Visualizations (optional advanced) ---
geopandas>=0.19.0
shapely>=2.0.1